requirements.txt
blinker==1.9.0
click==8.1.8
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==24.2
SQLAlchemy==2.0.40
typing_extensions==4.13.1
Werkzeug==3.1.3
streamlit
google-generativeai
python-dotenv